requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.7
cached-property==1.5.1
cachetools==3.1.1
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
docker==3.7.3
docker-compose==1.24.1
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
google-api-core==1.13.0
google-auth==1.6.3
google-auth-oauthlib==0.4.0
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
httplib2==0.13.0
idna==2.6
jsonschema==2.6.0
jwcrypto==0.6.0
keyring==17.1.1
keyrings.alt==3.1.1
oauth2client==4.1.3
oauthlib==3.0.2
paramiko==2.6.0
pigpio==1.42
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycrypto==2.6.1
PyGObject==3.30.4
PyNaCl==1.3.0
pyserial==3.4
python-apt==1.8.4
python-jwt==3.2.4
pytz==2019.1
pyxdg==0.25
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
RPi.GPIO==0.6.5
rsa==4.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
texttable==0.9.1
urllib3==1.24.1
websocket-client==0.56.0
